Time complexity
Internally, a `StringBuilder` is a Char Array. `.append()` will iterate over the String passed as an argument and add each Char to the end of the Array. Therefore, calling `.append("al")` will be O(2), but we simplify this to O(1). Since big O is also not concerned with non-dominant terms, we can drop the O(M) from calling `sb.toString()`, which will leave us with a linear running time